[ICASSP 2023] Dev update

+++ nnsvs/bin/prepare_voc_features.py
@@ -2,7 +2,7 @@
 """
 import os
 from concurrent.futures import ProcessPoolExecutor
-from os.path import exists, join
+from os.path import exists, islink, join
 
 import hydra
 import numpy as np
@@ -26,7 +26,6 @@ def _prepare_voc_features(
     in_wave_path = join(in_dir, utt_id + "-wave.npy")
     assert exists(in_wave_path)
 
-    assert np.any(has_dynamic_features)
     streams = get_static_features(
         feats.reshape(1, -1, feats.shape[-1]),
         num_windows,
@@ -39,9 +38,12 @@ def _prepare_voc_features(
 
     # NOTE: even if the number of streams are larger than 4, we only use the first 4 streams
     # for training neural vocoders
-    assert len(streams) >= 4
-    mgc, lf0, vuv, bap = streams[0], streams[1], streams[2], streams[3]
-    voc_feats = np.hstack((mgc, lf0, vuv, bap)).astype(np.float32)
+    if len(streams) >= 4:
+        mgc, lf0, vuv, bap = streams[0], streams[1], streams[2], streams[3]
+        voc_feats = np.hstack((mgc, lf0, vuv, bap)).astype(np.float32)
+    elif len(streams) == 3:
+        mel, lf0, vuv = streams[0], streams[1], streams[2]
+        voc_feats = np.hstack((mel, lf0, vuv)).astype(np.float32)
 
     voc_feats_path = join(out_dir, utt_id + "-feats.npy")
     np.save(voc_feats_path, voc_feats, allow_pickle=False)
@@ -49,7 +51,7 @@ def _prepare_voc_features(
     # NOTE: To train vocoders with https://github.com/kan-bayashi/ParallelWaveGAN
     # target waveform needs to be created in the same directory as the vocoder input features.
     save_wave_path = join(out_dir, utt_id + "-wave.npy")
-    if not exists(save_wave_path):
+    if (not exists(save_wave_path)) and (not islink(save_wave_path)):
         os.symlink(join(in_dir, utt_id + "-wave.npy"), save_wave_path)
 
 
@@ -67,12 +69,17 @@ def my_app(config: DictConfig) -> None:
     if config.acoustic.relative_f0:
         raise ValueError("Relative F0 is not supported")
 
-    stream_sizes, has_dynamic_features = get_world_stream_info(
-        config.acoustic.sample_rate,
-        config.acoustic.mgc_order,
-        config.acoustic.num_windows,
-        config.acoustic.vibrato_mode,
-    )
+    if config.acoustic.feature_type == "world":
+        stream_sizes = get_world_stream_info(
+            config.acoustic.sample_rate,
+            config.acoustic.mgc_order,
+            config.acoustic.num_windows,
+            config.acoustic.vibrato_mode,
+        )
+    elif config.acoustic.feature_type == "melf0":
+        stream_sizes = [80, 1, 1]
+    else:
+        raise ValueError(f"Unknown feature type: {config.acoustic.feature_type}")
 
     os.makedirs(out_dir, exist_ok=True)
     with ProcessPoolExecutor(max_workers=config.max_workers) as executor:
@@ -84,7 +91,7 @@ def my_app(config: DictConfig) -> None:
                 utt_id,
                 config.acoustic.num_windows,
                 stream_sizes,
-                has_dynamic_features,
+                config.acoustic.dynamic_features_flags,
             )
             for utt_id in utt_ids
         ]
